Move export products to tasks file
